requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djoser==2.1.0
Pillow==9.5.0
django_debug_toolbar==4.0.0